refactoring galore
