+++ ulmo/llc/uniform.py
@@ -161,50 +161,3 @@ def extract_preproc_for_analysis(llc_table, preproc_root='llc_std',
 
     # Return
     return llc_table
-
-def main(flg):
-    if flg== 'all':
-        flg= np.sum(np.array([2 ** ii for ii in range(25)]))
-    else:
-        flg= int(flg)
-
-    # Testing
-    if flg & (2**0):  
-        # Generate the extraction file
-        resol = 0.5  # deg
-        #outfile = os.path.join(os.getenv('SST_OOD'), 'LLC', 'Tables', 'test_uniform_0.5.csv')
-        outfile = 's3://llc/Tables/test_uniform_0.5.csv'
-        if os.path.isfile(outfile):
-            print("{} exists.  Am loading it.".format(outfile))
-            llc_table = llc_io.load_llc_table(outfile)
-        else:
-            llc_table = extract.uniform_coords(resol=resol, field_size=(64,64),
-                                               outfile=outfile)
-        # Plot
-        if False:
-            extract.plot_extraction(llc_table, s=1, resol=resol)
-
-        # Extract 6 days across the full range;  ends of months
-        dti = pandas.date_range('2011-09-01', periods=6, freq='2M')
-        llc_table = extract.add_days(llc_table, dti, outfile=outfile)
-
-
-        # Giddy up (will take a bit of memory!)
-        pp_file = os.path.join(os.getenv('SST_OOD'), 'LLC', 'PreProc', 
-                               'LLC_uniform_preproc_test.h5')
-        llc_table = extract_preproc_for_analysis(llc_table, outfile=pp_file)
-        # Final write
-        llc_io.write_llc_table(llc_table, outfile)
-
-
-# Command line execution
-if __name__ == '__main__':
-    import sys
-
-    if len(sys.argv) == 1:
-        flg = 0
-        #flg += 2 ** 0  # outliers
-    else:
-        flg = sys.argv[1]
-
-    main(flg)
