+++ ulmo/runs/LLC/uniform/llc_test_extract_pp.yaml
@@ -1,0 +1,66 @@
+apiVersion: batch/v1
+kind: Job
+metadata:
+  name: xavier-llc-extract-test
+spec:
+  backoffLimit: 0
+  template:
+    spec:
+      containers:
+      - name: container
+        image: localhost:30081/profxj/ulmo_nvidia:latest  # UPDATE
+        imagePullPolicy: Always
+        resources:
+          requests:
+            cpu: "10"
+            memory: "32Gi"
+            ephemeral-storage: 32Gi
+          limits:
+            cpu: "16"
+            memory: "128Gi"
+            ephemeral-storage: 150Gi
+            #nvidia.com/gpu:  "1"  # See docs to exlude certain types
+            # https://ucsd-prp.gitlab.io/userdocs/running/gpu-pods/
+        command: ["/bin/bash", "-c"]
+        args:
+          - cd ulmo; 
+            git checkout models; 
+            git pull; 
+            python setup.py develop;
+            cd ../tmp; 
+            mkdir foo; 
+            cd foo; 
+            mkdir PreProc;   
+            cp /ulmo/runs/LLC/llc_evaluate_test.py .;
+            python -u /ulmo/runs/LLC/llc_test_extract.py;
+            cd ..;
+            aws --endpoint http://rook-ceph-rgw-nautiluss3.rook s3 cp foo s3://llc --recursive --force;
+        env:
+          - name: "ENDPOINT_URL"
+            value: "http://rook-ceph-rgw-nautiluss3.rook"
+          - name: "S3_ENDPOINT"
+            value: "rook-ceph-rgw-nautiluss3.rook"
+        volumeMounts:
+          - name: prp-s3-credentials
+            mountPath: "/root/.aws/credentials"
+            subPath: "credentials"
+          - name: ephemeral
+            mountPath: "/tmp"
+          - name: "dshm"
+            mountPath: "/dev/shm"
+      nodeSelector:
+        nautilus.io/disktype: nvme
+        gpu-type: "1080Ti"  # titan-xp
+      restartPolicy: Never
+      volumes:
+        # Secrets file for nautilus s3 credentials .aws/credentials and .s3cfg
+        - name: prp-s3-credentials
+          secret:
+            secretName: prp-s3-credentials
+        # Shared memory (necessary for Python's multiprocessing.shared_memory module to work)
+        - name: dshm
+          emptyDir:
+            medium: Memory
+        # Ephemeral storage
+        - name: ephemeral
+          emptyDir: {}
