+++ ulmo/runs/Latents/modis_l2_latents.yaml
